wip: phases as derivations
The basic idea, and goal, being that we can take a derivation, and through a series of overrides, run each
phase in a separate derivation. This would make interactive development a little easier, as troubleshooting
a failure during installPhase doesn't require re-running the buildPhase.

As an example, with

```console
$ cat ~/.config/nixpkgs/config.nix
{
  allowUnfree = true;
  cudaSupport = true;
  cudaCapabilities = [ "8.9" ];
}
```

build with

```console
nix build --impure -L .#magma.phases.installed
```

Make a change to the postInstall of the magma derivation in pkgs/development/libraries/science/math/magma/generic.nix.

Run the build again.

See how it is able to re-use the previous stages?